fix a few circular dependencies
