+++ src/constants/index.ts
@@ -1,1 +1,11 @@
 export const DEFAULT_ENTRY_POINT = '0x0000000071727De22E5E9d8BAf0edAc6f37da032'
+
+/**
+ * Name of the native tracer that is equivalent to the standard javascript
+ */
+export const bundlerNativeTracerName = 'bundlerCollectorTracer'
+
+/**
+ * Name of the prestate tracer
+ */
+export const prestateTracerName = 'prestateTracer'
