+++ src/mempool/index.ts
@@ -1,2 +1,1 @@
 export * from './mempool-manager.js'
-export * from './mempool.types.js'
