+++ src/metrics/index.ts
@@ -1,2 +1,1 @@
 export * from './metrics-tracker.js'
-export * from './metrics.types.js'
