+++ src/reputation/index.ts
@@ -1,2 +1,1 @@
 export * from './reputation-manager.js'
-export * from './reputation.types.js'
