+++ src/rpc/apis/Web3.ts
@@ -1,13 +1,4 @@
-export type Web3APIMethodMapping = {
-  web3_clientVersion: {
-    params: []
-    return: string
-  }
-}
-
-export type Web3API = {
-  clientVersion(): string
-}
+import { Web3API } from '../../types/index.js'
 
 export const createWeb3API = (version: string): Web3API => {
   return {
