+++ src/rpc/index.ts
@@ -1,4 +1,3 @@
 export * from './rpc-server.js'
-export * from './rpc.types.js'
 export * from './bundler-handler-registry.js'
 export * from './apis/index.js'
