+++ src/rpc/rpc-server.ts
@@ -11,10 +11,9 @@ import type {
   RpcHandler,
   JsonRpcErrorResponse,
   JsonRpcResponse,
-} from './rpc.types.js'
+  RpcError,
+} from '../types/index.js'
 import { createRpcHandler } from './rpc-handler.js'
-import { Either } from '../monad/index.js'
-import { RpcError } from '../utils/index.js'
 
 const createApp = (rpc: RpcHandler): express.Application => {
   const app = express()
@@ -41,32 +40,34 @@ const createApp = (rpc: RpcHandler): express.Application => {
     Logger.debug(
       `---> Handling valid request for ${request.method} with requestId(${request.id})`,
     )
-    const result = await rpc.doHandleRequest(request).catch((error) => {
+    try {
+      const result = await rpc.doHandleRequest(request)
+      result.fold(
+        (error: RpcError) =>
+          res.json({
+            ...errorRes,
+            error: {
+              code: error.code,
+              message: error.message,
+              data: error.data,
+            },
+          }),
+        (response: JsonRpcResponse) => res.json(response),
+      )
+    } catch (error: any) {
       Logger.error(
         { error: error.message },
         `Unknown error handling method requestId(${request.id})`,
       )
-      return Either.Left<RpcError, JsonRpcResponse>(
-        new RpcError(
-          error.message ? error.message : errorRes.error.message,
-          error.code ? error.code : errorRes.error.code,
-          error.data,
-        ),
-      )
-    })
-
-    result.fold(
-      (error: RpcError) =>
-        res.json({
-          ...errorRes,
-          error: {
-            code: error.code,
-            message: error.message,
-            data: error.data,
-          },
-        }),
-      (response: JsonRpcResponse) => res.json(response),
-    )
+      res.json({
+        ...errorRes,
+        error: {
+          code: error.code ? error.code : errorRes.error.code,
+          message: error.message ? error.message : errorRes.error.message,
+          data: error.data,
+        },
+      })
+    }
   })
 
   return app
