+++ src/signer/signerService.ts
@@ -1,19 +1,6 @@
-import { Wallet } from 'ethers'
 import { Logger } from '../logger/index.js'
-import { BundleTxs } from '../state/index.js'
+import { BundleTxs, SignerService } from '../types/index.js'
 import { ProviderService } from '../provider/index.js'
-
-export type BundlerSignerWallets = Record<number, Wallet>
-
-export type SignerService = {
-  /**
-   * Finds the first signer without a pending bundle transaction and returns the index.
-   *
-   * @param bundleTxs - Record of bundleTxs.
-   * @returns - Return the index of the available signer, -1 if all are busy, or 0 immediately if bundleTxs is empty.
-   */
-  getReadySigner(bundleTxs: BundleTxs): Promise<number>
-}
 
 // Utility function to check if a transaction is pending
 const isTransactionPending = async (
