+++ src/sim/index.ts
@@ -1,3 +1,1 @@
 export * from './sim.js'
-export * from './gethTracer.types.js'
-export * from './sim-types.js'
