+++ src/state/index.ts
@@ -1,2 +1,1 @@
-export * from './state.types.js'
 export * from './state.js'
