+++ src/utils/rpc.utils.ts
@@ -1,4 +1,5 @@
 import { toBeHex } from 'ethers'
+import { RpcError } from '../types/index.js'
 
 export const requireCond = (
   cond: boolean,
@@ -8,26 +9,6 @@ export const requireCond = (
 ): void => {
   if (!cond) {
     throw new RpcError(msg, code, data)
-  }
-}
-
-export class RpcError extends Error {
-  // error codes from: https://eips.ethereum.org/EIPS/eip-1474
-  constructor(
-    msg: string,
-    readonly code: number,
-    readonly data: any = undefined,
-  ) {
-    super(msg)
-  }
-}
-
-export class NetworkCallError extends Error {
-  constructor(
-    msg: string,
-    readonly payload: any,
-  ) {
-    super(msg)
   }
 }
