+++ src/utils/validation.utils.ts
@@ -1,7 +1,6 @@
 import { BigNumberish, ethers, zeroPadValue, dataSlice, toBeHex } from 'ethers'
 
-import { UserOperation } from '../types/index.js'
-import { ValidationErrors } from '../validation/index.js'
+import { UserOperation, ValidationErrors } from '../types/index.js'
 
 import { requireCond } from './rpc.utils.js'
