+++ src/validation/index.ts
@@ -1,2 +1,1 @@
 export * from './validation-service.js'
-export * from './validation.types.js'
