+++ vitest.config.ts
@@ -4,7 +4,7 @@ import { defineConfig } from 'vitest/config'
 // https://github.com/NomicFoundation/hardhat/issues/3385
 export default defineConfig({
   test: {
-    include: ['src/**/*.test.ts'],
+    include: ['test/*.test.ts'],
     exclude: [],
     setupFiles: './vitest.setup.ts',
     environment: 'node',
